+++ TDEngine2/plugins/D3D12GraphicsContext/source/CD3D12GraphicsContext.cpp
@@ -317,31 +317,12 @@ namespace TDEngine2
 			void BindRasterizerState(TRasterizerStateId rasterizerStateId) override;
 
 			/*!
-				\brief The method binds a given render target object to rendering pipeline
-
-				\param[in] slot An index of the slot into which the render target will be bound
-
-				\param[in] targetHandle Handle to texture object that's created as a render target
+				\brief The pair of methods BeginRenderPass/EndRenderPass are intended to replace separate calls of BindRenderTarget/BindDepthBufferTarget/SetDepthBufferEnabled and others
+				to provide single point configuration of all targets that participate in rendering. The concept should be familiar for ones who worked with OpenGL/Vulkan GAPIs
 			*/
 
-			void BindRenderTarget(U8 slot, TTextureHandleId targetHandle) override;
-
-			/*!
-				\brief The method binds a given depth buffer to rendering pipeline
-
-				\param[in] targetHandle Handle to texture object that's created as a depth buffer
-				\param[in] disableRTWrite A flag determines whether the write to RT should be enabled or not
-			*/
-
-			void BindDepthBufferTarget(TTextureHandleId targetHandle, bool disableRTWrite = false) override;
-
-			/*!
-				\brief The method disables or enables a depth buffer usage
-
-				\param[in] value If true the depth buffer will be used, false turns off it
-			*/
-
-			void SetDepthBufferEnabled(bool value) override;
+			TDE2_API E_RESULT_CODE BeginRenderPass(const TFramebufferInfo& framebufferInfo) override;
+			TDE2_API E_RESULT_CODE EndRenderPass() override;
 
 			/*!
 				\brief The method returns an object that contains internal handlers that are used by the system.
@@ -1080,16 +1061,14 @@ namespace TDEngine2
 	{
 	}
 
-	void CD3D12GraphicsContext::BindRenderTarget(U8 slot, TTextureHandleId targetHandle)
+	E_RESULT_CODE CD3D12GraphicsContext::BeginRenderPass(const TFramebufferInfo& framebufferInfo)
 	{
+		return RC_NOT_IMPLEMENTED_YET;
 	}
 
-	void CD3D12GraphicsContext::BindDepthBufferTarget(TTextureHandleId targetHandle, bool disableRTWrite)
+	E_RESULT_CODE CD3D12GraphicsContext::EndRenderPass()
 	{
-	}
-
-	void CD3D12GraphicsContext::SetDepthBufferEnabled(bool value)
-	{
+		return RC_NOT_IMPLEMENTED_YET;
 	}
 
 	const TGraphicsCtxInternalData& CD3D12GraphicsContext::GetInternalData() const